// app/api/models/response/verify_response.ts
export interface VerifyResponse {
    id: number;
    idAnimalAi: number;
    idAnimal: number;
    nameAnimalAi: string;
    nameAnimal: string;
    urlAi: string;
    url: string;
    date: Date;
    statusDescription: string;
    status: string;
    categoryId: number;
    analistId: number;
    analistName: string;
    idAnimalUser: number;
    nameAnimalUser: string;
}